refactor(sysmon): move sysmon types to src/workspaces/sysmon/
